+++ tests/integration/test_get_location.py
@@ -1,0 +1,13 @@
+from ebird.api import get_location
+
+
+def test_get_hotspot(api_token):
+    identifier = "L901738"
+    details = get_location(api_token, identifier)
+    assert details["locId"] == identifier
+
+
+def test_get_private_location(api_token):
+    identifier = "L8978128"
+    details = get_location(api_token, identifier)
+    assert details["locId"] == identifier